requirements.txt.txt
Flask
tensorflow
numpy
pandas
matplotlib
seaborn
scikit-learn
Pillow